requirements.txt
pywin32>=227
PyYAML~=5.4.1
ahkunwrapped~=2.0.2
